refactor: move proposal action models to proposal actions module
