+++ src/proposals/models/proposal.model.ts
@@ -14,7 +14,7 @@ import { Image } from "../../images/models/image.model";
 import { User } from "../../users/models/user.model";
 import { Vote } from "../../votes/models/vote.model";
 import { ProposalStages } from "../proposals.constants";
-import { ProposalAction } from "./proposal-action.model";
+import { ProposalAction } from "../proposal-actions/models/proposal-action.model";
 
 @Entity()
 @ObjectType()
