+++ src/proposals/proposals.resolver.ts
@@ -17,7 +17,7 @@ import { User } from "../users/models/user.model";
 import { Vote } from "../votes/models/vote.model";
 import { CreateProposalInput } from "./models/create-proposal.input";
 import { CreateProposalPayload } from "./models/create-proposal.payload";
-import { ProposalAction } from "./models/proposal-action.model";
+import { ProposalAction } from "./proposal-actions/models/proposal-action.model";
 import { Proposal } from "./models/proposal.model";
 import { UpdateProposalInput } from "./models/update-proposal.input";
 import { UpdateProposalPayload } from "./models/update-proposal.payload";
